requirements.txt
# 문서 텍스트 추출을 위한 패키지들
PyMuPDF>=1.26.0          # PDF 파일 추출
python-docx>=1.2.0       # DOCX 파일 추출
python-pptx>=1.0.0       # PPTX 파일 추출
pandas>=2.3.0            # Excel 파일 추출
openpyxl>=3.1.0          # XLSX 파일 지원
xlrd>=2.0.0              # XLS 파일 지원

# 향후 파인튜닝을 위한 패키지들 (예정)
# torch>=2.0.0
# transformers>=4.30.0
# peft>=0.4.0
# trl>=0.7.0
# accelerate>=0.20.0 